rails generate rspec:install
Next, we need to configure Rails to use RSpec in place of Test::Unit.
This can be accomplished with rails generate rspec:install:

$ rails generate rspec:install

If your system complains about the lack of a JavaScript runtime, visit
the execjs page at GitHub for a list of possibilities. I particularly
recommend installing Node.js.