3.6.3 Speeding up tests with Spork
When running bundle exec rspec, you may have noticed that it takes
several seconds just to start running the tests, but once they start
running they finish quickly. This is because each time RSpec runs the
tests it has to reload the entire Rails environment. The Spork test
server13 aims to solve this problem. Spork loads the environment once,
and then maintains a pool of processes for running future tests. Spork
is particularly useful when combined with Guard (Section 3.6.2).