5.1.2 Bootstrap and custom CSS
In Section 5.1.1, we associated many of the HTML elements with CSS
classes, which gives us considerable flexibility in constructing a
layout based on CSS. As noted in Section 5.1.1, many of these classes
are specific to Bootstrap, a framework from Twitter that makes it easy
to add nice web design and user interface elements to an HTML5
application. In this section, we’ll combine Bootstrap with some custom
CSS rules to start adding some style to the sample application.