8.2 Signin Success
Having handled a failed signin, we now need to actually sign a user in.
Getting there will require some of the most challenging Ruby programming
so far in this tutorial, so hang in there through the end and be
prepared for a little heavy lifting. Happily, the first step is
easy—completing the Sessions controller create action is a snap.
Unfortunately, it’s also a cheat.