9.1 Updating Users
The pattern for editing user information closely parallels that for
creating new users (Chapter 7). Instead of a new action rendering a view
for new users, we have an edit action rendering a view to edit users;
instead of create responding to a POST request, we have an update action
responding to a PATCH request (Box 3.3). The biggest difference is that,
while anyone can sign up, only the current user should be able to update
their information. This means that we need to enforce access control so
that only authorized users can edit and update; the authentication
machinery from Chapter 8 will allow us to use a before filter to ensure
that this is the case.