9.2.3 Friendly Forwarding
Our site authorization is complete as written, but there is one minor
blemish: when users try to access a protected page, they are currently
redirected to their profile pages regardless of where they were trying
to go. In other words, if a non-logged-in user tries to visit the edit
page, after signing in the user will be redirected to /users/1 instead
of /users/1/edit. It would be much friendlier to redirect them to their
intended destination instead.

To test for such “friendly forwarding”, we first visit the user edit
page, which redirects to the signin page. We then enter valid signin
information and click the “Sign in” button. The resulting page, which by
default is the user’s profile, should in this case be the “Edit user”
page. The test for this sequence appears in Listing 9.16.