9.3 Showing all Users
In this section, we’ll add the penultimate user action, the index
action, which is designed to display all the users instead of just one.
Along the way, we’ll learn about populating the database with sample
users and paginating the user output so that the index page can scale up
to display a potentially large number of users. A mockup of the
result—users, pagination links, and a “Users” navigation link—appears in
Figure 9.7.6 In Section 9.4, we’ll add an administrative interface to
the user index so that (presumably troublesome) users can be destroyed.