9.3.2 Sample Users
In this section, we’ll give our lonely sample user some company. Of
course, to create enough users to make a decent user index, we could use
our web browser to visit the signup page and make the new users one by
one, but far a better solution is to use Ruby (and Rake) to make the
users for us.