10.1 A Micropost Model
We begin the Microposts resource by creating a Micropost model, which
captures the essential characteristics of microposts. What follows
builds on the work from Section 2.3; as with the model in that section,
our new Micropost model will include data validations and an association
with the User model. Unlike that model, the present Micropost model will
be fully tested, and will also have a default ordering and automatic
destruction if its parent user is destroyed.