10.2 Showing Microposts
Although we don’t yet have a way to create microposts through the
web—that comes in Section 10.3.2—that won’t stop us from displaying them
(and testing that display). Following Twitter’s lead, we’ll plan to
display a user’s microposts not on a separate microposts index page, but
rather directly on the user show page itself, as mocked up in Figure
10.4. We’ll start with fairly simple ERb templates for adding a
micropost display to the user profile, and then we’ll add microposts to
the sample data populator from Section 9.3.2 so that we have something
to display.