11.2 A web interface for following users
In the introduction to this chapter, we saw a preview of the page flow
for user following. In this section, we will implement the basic
interface and following/unfollowing functionality shown in those
mockups. We will also make separate pages to show the user following and
followers arrays. In Section 11.3, we’ll complete our sample application
by adding the user’s status feed.