11.4 Conclusion
With the addition of the status feed, we’ve finished the core sample
application for the Ruby on Rails Tutorial. This application includes
examples of all the major features of Rails, including models, views,
controllers, templates, partials, filters, validations, callbacks,
has_many/belongs_to and has_many through associations, security,
testing, and deployment. Despite this impressive list, there is still
much to learn about Rails. As a first step in this process, this section
contains some suggested extensions to the core application, as well as
suggestions for further learning.